Changed single searches names
